Make DBSummariser great again
It broke when I added wind. I'm not sure if that's what broke it, or if
it was broken all the time.
Check the change in: db/src/DBSummariser/Summariser.java

Also, this change breaks old DB files, as temp is not +1000 before it is
stored, and temp can not not be below -100.0C.